requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
boto3==1.28.0
python-dotenv==1.0.0
cryptography==46.0.7
DBUtils==1.3
tenacity==8.2.3
werkzeug==3.0.0
